pyproject.toml
[project]
name = "moodle"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "playwright>=1.49.1",
    "python-dotenv>=1.0.1",
    "python-telegram-bot>=21.10",
    "requests>=2.32.3",
]
